refactor: move embedder to backbone
